+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-16T07:25:50.726Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-16T10:40:20.235Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")