+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7NYJMC0S3K25AKTYQGJM82J"
+        "01K7P9PQVTVP4YP27NMMJM8Q7Q"
     )
     r"""The id of the resource"""