+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-16T07:25:54.018Z")
+    created: Optional[datetime] = parse_datetime("2025-10-16T10:40:23.258Z")
     r"""The date and time the resource was created"""