+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7NYJMBYBZ8TEWVW0YFJ36A9"
+        "01K7P9PQVRCPK5MVQN5NERYF5T"
     )
     r"""The id of the resource"""