+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7NYJMBZBFTABQ5TVYK6CPJT"
+        "01K7P9PQVSJ222H93WF5C8ZR9R"
     )
     r"""The id of the resource"""