+++ packages/orq-rc/src/orq_ai_sdk/models/updatebudgetop.py
@@ -208,5 +208,5 @@ class UpdateBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-16T07:25:50.726Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-16T10:40:20.235Z")
     r"""The date and time the resource was last updated"""