+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7NYJMC18NNWV1NJBTCRWCFH"
+        "01K7P9PQVVTMR4C8QT80HRJ9G0"
     )
     r"""The id of the resource"""