+++ packages/orq-rc/RELEASES.md
@@ -3369,3 +3369,13 @@ Based on:
 - [python v3.14.0-rc.32] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.32] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.32 - packages/orq-rc
+
+## 2025-10-16 10:42:01
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.33] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.33] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.33 - packages/orq-rc